s: tipi rigenerati (e file rinominati)
