+++ shared/src/types/typesFixes.ts
@@ -1,10 +1,10 @@
-import { UsersPermissionsMe } from "./strapi";
+import { UsersPermissionsMe } from "./types";
 
 export type ID<T> = T & { id: string };
 
 export type Comp<T> = Partial<T> & { __component: string };
 
 export interface LoginResponse {
-    jwt: string;
-    user: UsersPermissionsMe;
+	jwt?: string;
+	user: UsersPermissionsMe;
 }
